+++ packages/react/src/index.ts
@@ -71,4 +71,4 @@ export {
 /**
  * Package version
  */
-export const VERSION = '2.0.0';
+export const VERSION = '2.0.1';