+++ packages/docs/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @create-markdown/docs
 
+## 1.0.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @create-markdown/core@2.0.1
+  - @create-markdown/mdx@1.0.1
+
 ## 1.0.0
 
 ### Major Changes